Add totally hacked-together Markdown completer
This is gross, not idiomatic Python, and probably a horrible abuse of the
ycmd development model, but it is also incredibly useful to me.

This allows me to autocomplete wiki and blog links in my Markdown documents
on typing `](`. It connects to the local Masochist Redis server to get the list
of articles, and it relies on a `.ycm_extra_conf.py` file that is specific to
the Masochist content repo in order to get the metadata necessary to do the
Redis lookup (ie. a key prefix and a cache breaker). I could have hard-coded
this in the completer but that felt wrong. So instead I've done this ugly hoop
jumping. I feel ashamed of how crap this is, but it works, so I am keeping it.

Feedback welcome if anybody reading this knows better.